fix typos and refactor code
